Make mpi_util for defining MPI point type
