+++ Makefile
@@ -6,11 +6,12 @@ serial: src/serial.c src/k_clustering.c $(CORE_FILES) $(HEADER_FILES)
 	gcc -Wall src/serial.c src/k_clustering.c $(CORE_FILES) -o output/serial
 shared_cpu: src/shared_cpu.c src/shared_cpu_k_clustering.c $(CORE_FILES) $(HEADER_FILES)
 	gcc -Wall -fopenmp src/shared_cpu.c src/shared_cpu_k_clustering.c $(CORE_FILES) -o output/shared_cpu
-distributed_cpu: src/distributed_cpu.c src/distributed_cpu_k_clustering.c $(CORE_FILES) $(HEADER_FILES)
-	mpicc -g -Wall -std=c99 -o output/distributed_cpu src/distributed_cpu.c src/distributed_cpu_k_clustering.c $(CORE_FILES)
+distributed_cpu: src/distributed_cpu.c src/distributed_cpu_k_clustering.c $(CORE_FILES) $(HEADER_FILES) src/mpi_util.c src/mpi_util.h
+	mpicc -g -Wall -std=c99 -o output/distributed_cpu src/distributed_cpu.c src/distributed_cpu_k_clustering.c $(CORE_FILES) src/mpi_util.c
 shared_gpu: src/shared_gpu.c src/shared_gpu_k_clustering.c src/kernel.cu $(CORE_FILES) $(HEADER_FILES)
 	nvcc src/shared_gpu.c src/shared_gpu_k_clustering.c $(CORE_FILES) src/kernel.cu -o output/shared_gpu
 distributed_gpu: src/shared_gpu.c src/shared_gpu_k_clustering.c src/kernel.cu $(CORE_FILES) $(HEADER_FILES)
+	mpicc -g -Wall -std=c99 -c src/mpi_util.c -o output/mpi_util.o
 	mpicc -g -Wall -std=c99 -c src/point.c -o output/point.o
 	mpicc -g -Wall -std=c99 -c src/csv.c -o output/csv.o
 	mpicc -g -Wall -std=c99 -c src/distributed_gpu_k_clustering.c -o output/distributed_gpu_k_clustering.o
