+++ src/distributed_cpu_k_clustering.h
@@ -7,4 +7,3 @@ void k_means_clustering(Point* points,
 			int points_size,
 			int my_rank,
 			int comm_sz);
-MPI_Datatype create_point_datatype();
