+++ src/mpi_util.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <mpi.h>
+
+MPI_Datatype create_point_datatype();
